Final v1.1.0-alpha commit, merge to master
