+++ src/nordland.py
@@ -1,25 +1,3 @@
-#MIT License
-
-#Copyright (c) 2023 Adam Hines, Peter G Stratton, Michael Milford, Tobias Fischer
-
-#Permission is hereby granted, free of charge, to any person obtaining a copy
-#of this software and associated documentation files (the "Software"), to deal
-#in the Software without restriction, including without limitation the rights
-#to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-#copies of the Software, and to permit persons to whom the Software is
-#furnished to do so, subject to the following conditions:
-
-#The above copyright notice and this permission notice shall be included in all
-#copies or substantial portions of the Software.
-
-#THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-#IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-#FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-#AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-#LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-#OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-#SOFTWARE.
-
 '''
 Imports
 '''
@@ -37,12 +15,23 @@ def atoi(text):
 def natural_keys(text):
     return [ atoi(c) for c in re.split(r'(\d+)', text) ]
 
-basePath = "/home/adam/data/VPRTempo_training/"
+# set the base path to the location of the downloaded Nordland datasets
+basePath = ''
+assert(os.path.isdir(basePath)),"Please set the basePath to the location of the downloaded Nordland datasets"
+
+# define the subfolders of the Nordland datasets
 subPath = ["spring_images_train/section1/","spring_images_train/section2/",
            "fall_images_train/section1/","fall_images_train/section2/",
            "winter_images_train/section1/","winter_images_train/section2/",
            "summer_images_train/section1/","summer_images_train/section2/"]
-outPath = ["training_data/spring/","training_data/fall/","training_data/winter/","testing_data/"]
+
+# set the desired output folder for unzipping and organization
+outDir = ''
+assert(os.path.isdir(outDir)),"Please set the outDir to the desired output location for unzipping the Nordland datasets"
+
+# define output paths for the data
+outPath = [os.path.join(outDir,"spring/"),os.path.join(outDir,"fall/"),
+           os.path.join(outDir,"winter/"),os.path.join(outDir,"summer/")]
 
 # check for existence of the zip folders, throw exception if missing
 zipNames = ["spring_images_train.zip","fall_images_train.zip",
@@ -54,6 +43,7 @@ for n in zipNames:
 # check if nordland data folders have already been unzipped
 zip_flag = []
 for n, ndx in enumerate(range(0,len(subPath),2)):
+    print('Unzipping '+zipNames[n])
     if os.path.exists(basePath+subPath[ndx]):
         # check if the folder contains any files
         file_lst = os.listdir(basePath+subPath[ndx])
@@ -78,18 +68,15 @@ for n in range(0,len(subPath)):
     tempPaths = [basePath+subPath[n]+s for s in tempPaths]
     imgPaths = imgPaths + tempPaths
 
-# if training and testing folders already exist, delete them
-if os.path.exists(basePath+'training_data/'):
-    shutil.rmtree(basePath+'training_data/')
-    print('Removed pre-existing training data folder')
-if os.path.exists(basePath+'testing_data/'):
-    shutil.rmtree(basePath+'testing_data/')
-    print('Remove pre-existing testing data folder')
-    
-# rename and move the training data
-os.mkdir(basePath+"training_data/")
+# if output folders already exist, delete them
 for n in outPath:
-    os.mkdir(basePath+n)
+    if os.path.exists(n):
+        shutil.rmtree(n)
+        print('Removed pre-existing output folder')
+    
+# rename and move the training data to match the nordland_imageNames.txt file
+for n in outPath:
+    os.mkdir(n)
 for n, filename in enumerate(imgPaths):
     base = os.path.basename(filename)
     split_base = os.path.splitext(base)
@@ -104,14 +91,22 @@ for n, filename in enumerate(imgPaths):
     else:
         my_dest = "images-"+split_base[0] + ".png"
     if "spring" in filename:
-        out = basePath + outPath[0]
+        out = outPath[0]
     elif "fall" in filename:
-        out = basePath + outPath[1]
+        out = outPath[1]
     elif "winter" in filename:
-        out = basePath + outPath[2]
+        out = outPath[2]
     else:
-        out = basePath + outPath[-1]
+        out = outPath[-1]
 
     fileDest = out + my_dest   
     os.rename(filename, fileDest)
-    
+
+# remove the empty folders
+for n, ndx in enumerate(subPath):
+    if n%2 == 0:
+        shutil.rmtree(basePath+ndx.replace('section1/',''))
+    else:
+        continue
+
+print('Finished unzipping and organizing Nordland dataset')
